Persist original request URL when setting location
This change persists the original request URL within the locations
module so that if a user has clicked a specific link but was required
to sign in and then choose a location, they will still be redirected
to the original URL they requested rather than the dashboard for the
chosen location.
