+++ app/locations/controllers.test.js
@@ -46,135 +46,4 @@ describe('Locations controllers', function() {
       expect(params.locations).to.deep.equal(sortBy(mockUserLocations, 'title'))
     })
   })
-
-  describe('#setLocation', function() {
-    let nextSpy
-
-    beforeEach(function() {
-      req = {
-        session: {},
-        params: {},
-      }
-      res = {
-        redirect: sinon.spy(),
-      }
-      nextSpy = sinon.spy()
-    })
-
-    context('when no locationId parameter supplied', function() {
-      beforeEach(function() {
-        controllers.setLocation(req, res, nextSpy)
-      })
-
-      it('should call next without args', function() {
-        expect(nextSpy).to.be.calledOnceWithExactly()
-      })
-
-      it('should not set currentLocation', function() {
-        expect(req.session).not.to.have.property('currentLocation')
-      })
-
-      it('should not redirect', function() {
-        expect(res.redirect).not.to.be.called
-      })
-    })
-
-    context('when locationId is not found in user locations', function() {
-      beforeEach(function() {
-        req.params.locationId = 'not_authorised'
-        req.userLocations = mockUserLocations
-
-        controllers.setLocation(req, res, nextSpy)
-      })
-
-      it('should call next without args', function() {
-        expect(nextSpy).to.be.calledOnceWithExactly()
-      })
-
-      it('should not set currentLocation', function() {
-        expect(req.session).not.to.have.property('currentLocation')
-      })
-
-      it('should not redirect', function() {
-        expect(res.redirect).not.to.be.called
-      })
-    })
-
-    context('when locationId is found in user locations', function() {
-      beforeEach(function() {
-        req.params.locationId = mockUserLocations[0].id
-        req.userLocations = mockUserLocations
-
-        controllers.setLocation(req, res, nextSpy)
-      })
-
-      it('should set currentLocation', function() {
-        expect(req.session).to.have.property('currentLocation')
-        expect(req.session.currentLocation).to.deep.equal(mockUserLocations[0])
-      })
-
-      it('should redirect to homepage', function() {
-        expect(res.redirect).to.be.calledOnceWithExactly('/')
-      })
-
-      it('should not call next', function() {
-        expect(nextSpy).not.to.be.called
-      })
-    })
-  })
-
-  describe('#setAllLocations', function() {
-    let nextSpy
-
-    beforeEach(function() {
-      req = {
-        session: {},
-      }
-      res = {
-        redirect: sinon.spy(),
-      }
-      nextSpy = sinon.spy()
-    })
-
-    context('when user does not have permission', function() {
-      beforeEach(function() {
-        controllers.setAllLocations(req, res, nextSpy)
-      })
-
-      it('should call next without args', function() {
-        expect(nextSpy).to.be.calledOnceWithExactly()
-      })
-
-      it('should not set currentLocation', function() {
-        expect(req.session).not.to.have.property('currentLocation')
-      })
-
-      it('should not redirect', function() {
-        expect(res.redirect).not.to.be.called
-      })
-    })
-
-    context('when locationId is found in user locations', function() {
-      beforeEach(function() {
-        req.session.user = {
-          permissions: ['moves:view:all'],
-        }
-
-        controllers.setAllLocations(req, res, nextSpy)
-      })
-
-      it('should set currentLocation', function() {
-        expect(req.session).to.have.property('currentLocation')
-        expect(req.session.currentLocation).to.equal(null)
-      })
-
-      it('should redirect to homepage', function() {
-        expect(res.redirect).to.be.calledOnceWithExactly('/')
-      })
-
-      it('should not call next', function() {
-        expect(nextSpy).not.to.be.called
-      })
-    })
-  })
 })
