+++ app/locations/index.js
@@ -2,15 +2,21 @@
 const router = require('express').Router()
 
 // Local dependencies
-const { locations, setAllLocations, setLocation } = require('./controllers')
-const { setUserLocations, checkLocationsLength } = require('./middleware')
+const { redirect } = require('../auth/controllers')
+const { locations } = require('./controllers')
+const {
+  setUserLocations,
+  checkLocationsLength,
+  setLocation,
+  setAllLocations,
+} = require('./middleware')
 
 router.use(setUserLocations)
 
 // Define routes
 router.get('/', checkLocationsLength, locations)
-router.get('/all', setAllLocations)
-router.get('/:locationId', setLocation)
+router.get('/all', setAllLocations, redirect)
+router.get('/:locationId', setLocation, redirect)
 
 // Export
 module.exports = {
