+++ app/locations/middleware.js
@@ -1,4 +1,4 @@
-const { get } = require('lodash')
+const { find, get } = require('lodash')
 
 function setUserLocations(req, res, next) {
   req.userLocations = get(req.session, 'user.locations', [])
@@ -13,7 +13,28 @@ function checkLocationsLength(req, res, next) {
   next()
 }
 
+function setLocation(req, res, next) {
+  const { locationId } = req.params
+  const location = find(req.userLocations, { id: locationId })
+
+  req.session.currentLocation = location
+  next()
+}
+
+function setAllLocations(req, res, next) {
+  const { permissions = [] } = req.session.user || {}
+
+  if (!permissions.includes('moves:view:all')) {
+    return next()
+  }
+
+  req.session.currentLocation = null
+  next()
+}
+
 module.exports = {
   setUserLocations,
   checkLocationsLength,
+  setLocation,
+  setAllLocations,
 }
